requirements.txt
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
asgiref==3.8.1
awscli==1.33.16
botocore==1.34.134
certifi==2024.6.2
cffi==1.17.0rc1
charset-normalizer==3.3.2
colorama==0.4.6
crcmod==1.7
cryptography==42.0.8
Django==5.1a1
docutils==0.16
idna==3.7
jmespath==1.0.1
mysqlclient==2.2.4
oss2==2.18.6
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
setuptools==70.1.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
wheel==0.43.0
